% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etas.R
\name{etas}
\alias{etas}
\alias{etas.default}
\alias{etas.anova}
\alias{etas.lm}
\title{Eta coefficients}
\usage{
etas(object, ...)

\method{etas}{default}(object, fac, pop_var = FALSE, ...)

\method{etas}{anova}(object, ...)

\method{etas}{lm}(object, ...)
}
\arguments{
\item{object}{the R object}

\item{...}{arguments passed to other methods}

\item{fac}{vector for conditioning variable}

\item{pop_var}{logical, whether to use population or sample variance in the
calculation}
}
\value{
Values of eta and partial eta coefficients.
}
\description{
Calculate Eta coefficients, known as "Correlation ratios". Squared value of
Eta has an interpretation in terms of the proportion of explained variance.
The interpretation follows from a problem of predicting values of \eqn{y}.
The value of the \eqn{\eta^2}{Eta^2} is the proportion by which the error of
predicting values of \eqn{y} is reduced by using the information provided by
\eqn{x}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default method expect vectors. The function requires
additional argument \code{fac} -- a vector of the same length as \code{object}. The
result is a value (vector of length 1) of the \eqn{\eta^2}{Eta^2} assuming
that we want to predict the values of \code{object} with the values of \code{fac}
using the so called "Type I regression of means". For two variables \eqn{y}
and \eqn{x} the \eqn{\eta}{Eta} is given by the formula:

\deqn{\eta^2 = ( D^2(y) - E[D^2(y|x)] ) / D^2(y)}{Eta^2 = ( D^2(y) - E[D^2(y|x)] ) / D^2(y)}

\item \code{anova}: For objects of class \code{anova} the function calculates the
Eta's and Partial Eta Squares for all effects in the given model. In this
setting the eta squares for the given effect are equal to:

\deqn{\frac{SS_{effect}}{SS_{total}}}{SSeffect / SStotal}

where \eqn{SS} are appropriate Sums of Squares.  The ``Partial Eta Squares''
for the given effect are equal to:

\deqn{\frac{SS_{effect}}{SS_{effect}+SS_{resid}}}{SSeffect / (SSeffect+SSresid)}

The function returns a data frame with columns \code{Eta} and \verb{Partial Eta} with
a row for every term in the model.

\item \code{lm}: For objects of class \code{lm} the function is applied on the
result of calling \code{\link[=anova]{anova()}}.
}}

\examples{
### Generate some data
x1 <- rnorm(50)
x2 <- rnorm(50)
y <- 5 + 2 * x1 + rnorm(50, 0, 2) + 3 * x2 + rnorm(50, 0, .5)

### Method for vectors
etas(y, rep(1:2, each = 25))

### Method for 'lm' which calls 'anova'
m <- lm(y ~ x1 + x2)
etas(m)
}
