% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbstats-package.R
\docType{package}
\name{mbstats-package}
\alias{mbstats}
\alias{mbstats-package}
\title{mbstats: Michal's Beautiful Selection of Totally Awesome and Trustworthy Statistical functions}
\description{
Unsorted collection of various statististical functions. Perhaps
    some of them are useful also to others.
}
\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (\href{https://orcid.org/0000-0001-7503-852X}{ORCID})

}
\keyword{internal}
