% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_to.R
\name{rescale_to}
\alias{rescale_to}
\title{Rescale vector to given mean and SD}
\usage{
rescale_to(x, mu, s, na.rm = FALSE)
}
\arguments{
\item{x}{numeric; values to be rescaled}

\item{mu, sigma}{numeric; target mean and standard deviation}

\item{na.rm}{logical; passed to \code{\link[=mean]{mean()}} and \code{\link[=sd]{sd()}}}
}
\value{
Numeric vector of the same length as \code{x} of the rescaled values.
}
\description{
Rescale vector to given mean and SD
}
\examples{
x <- rescale_to(rnorm(20), 100, 50)
mean(x); sd(x)
}
\seealso{
\code{\link[=scale]{scale()}} and \code{\link[scales:rescale]{scales::rescale()}} for rescaling to new
minimum and maximum.
}
